스프링부트2 20210816 1번째 커밋

+++ src/main/resources/static/js/app/index.js
@@ -11,6 +11,21 @@ var main = {
             _this.delete();
         })
     },
+    deleteFile : function() {
+        var fileId = $('#file_id').val();
+        $.ajax({
+            async: false,//게시물 등록시 첨부파일은 비동기에서 동기로 바꿔야지만, 업로드 후 게시물이 저장됩니다.
+            type: 'DELETE',
+            url: '/api/file_delete/' + fileId,
+            //dataType: 'text',
+            dataType: 'json',
+            contentType: 'application/json; charset=utf-8',
+        }).done(function(result){
+            alert('첨부파일 삭제 성공 ' +  result.success);
+        }).fail(function(error){
+            alert('첨부파일 삭제 실패 ' + JSON.stringify(error));
+        });
+    },
     saveFile : function() {
         var _this = this;
         var form = $('#form_posts')[0];
@@ -19,13 +34,21 @@ var main = {
             async: false,//게시물 등록시 첨부파일은 비동기에서 동기로 바꿔야지만, 업로드 후 게시물이 저장됩니다.
             type: "POST",
             enctype: 'multipart/form-data',
-            url: "/api/upload",
+            url: "/api/file_upload",
             data: formData,
             processData: false,//formData 를 QueryString 으로 변환하지 않는다.
             contentType: false,//multipart/form-data; boundary=----WebKitFormBoundaryzS65BXVvgmggfL5A
             cache: false,
             timeout: 600000,
             dataType: 'json',
+            beforeSend:function(){
+                //이미지 보여주기 처리
+                $('.wrap-loading').removeClass('display-none');
+            },
+            complete:function(){
+                //이미지 숨김 처리
+                $('.wrap-loading').removeClass('display-none');
+            },
             success: function (result) {
                 alert("첨부파일 OK : " + result);
                 $("#file_id").val(result);
@@ -72,9 +95,17 @@ var main = {
         });
     },
     update : function () {
+        var _this = this;
+        if($("#customFile").val() != "") {
+            if($("#file_id").val() != "") {// && $("#file_id").val() != undefined) {
+                _this.deleteFile();
+            }
+            _this.saveFile();
+        }
         var data = {
             title: $('#title').val(),
-            content: $('#content').val()
+            content: $('#content').val(),
+            fileId: $('#file_id').val()
         };
         var id = $('#id').val();
         $.ajax({
@@ -91,6 +122,13 @@ var main = {
         });
     },
     delete : function () {
+        if(!confirm("정말로 삭제 하기겠습니까?")) {
+            return;
+        }
+        var _this = this;
+        if($("#file_id").val() != "") {// && $("#file_id").val() != undefined) {
+            _this.deleteFile();
+        }
         var id = $('#id').val();
         $.ajax({
             type: 'DELETE',
